backend/requirements.txt
flask
flask-cors
python-whois
requests
python-dotenv
gunicorn
google-generativeai